chore(android): remove old arch support

+++ src/specs/NativeAirMapsModule.ts
@@ -1,6 +1,6 @@
 import type {TurboModule} from 'react-native';
 import {TurboModuleRegistry} from 'react-native';
-import type {Double} from 'react-native/Libraries/Types/CodegenTypes';
+import type {Double, Int32} from 'react-native/Libraries/Types/CodegenTypes';
 import type {Camera} from './NativeComponentMapView';
 import type {Address} from '../MapView.types';
 
@@ -24,13 +24,13 @@ export type Region = Readonly<
 export type MapBoundaries = {northEast: LatLng; southWest: LatLng};
 
 export interface Spec extends TurboModule {
-  getCamera(tag: Double): Promise<Camera>;
-  getMarkersFrames(tag: Double, onlyVisible: boolean): Promise<unknown>;
-  getMapBoundaries(tag: Double): Promise<MapBoundaries>;
-  takeSnapshot(tag: Double, config: string): Promise<string>;
-  getAddressFromCoordinates(tag: Double, coordinate: LatLng): Promise<Address>;
-  getPointForCoordinate(tag: Double, coordinate: LatLng): Promise<Point>;
-  getCoordinateForPoint(tag: Double, point: Point): Promise<LatLng>;
+  getCamera(tag: Int32): Promise<Camera>;
+  getMarkersFrames(tag: Int32, onlyVisible: boolean): Promise<unknown>;
+  getMapBoundaries(tag: Int32): Promise<MapBoundaries>;
+  takeSnapshot(tag: Int32, config: string): Promise<string>;
+  getAddressFromCoordinates(tag: Int32, coordinate: LatLng): Promise<Address>;
+  getPointForCoordinate(tag: Int32, coordinate: LatLng): Promise<Point>;
+  getCoordinateForPoint(tag: Int32, point: Point): Promise<LatLng>;
 }
 
 export default TurboModuleRegistry.getEnforcing<Spec>('RNMapsAirModule');
